Small fix for kmin bounds checking
Previously, the code would check if kmin was larger than the machine epsilon, instead of larger than zero. As it turns out, this doesn't really make sense, since the machine epsilon is generally a relative measure of accuracy, at least across the non-denormal range.

As a quick check reveals, even the logarithm of the smallest denormal number is still a very manageable number, so right now the code just checks that kmin isn't literally zero. Denormal numbers are still unwanted, but that's probably not up to this piece of code to avoid.