Make the SIMD implementation a little faster
Previously, the SIMD implementation would follow the original
implementation and do two multiplications -- of Ak and xi, and of k and
kappa -- in the inner loop. However, these multiplications can be easily
done ahead of time, which is the subject of this patch. This appears to
indeed provide a non-negligible speedup to the implementation.