Implement a new optimization of TD13 SIMD code
This change expands the TD13 SIMD implementation by adding a custom vectorized cosine to increase performance. It also obviates the need to build and install SLEEF for its cosine implementation.

This is the fastest I could make it without using floats and losing precision. I will add/update the documentation in future commits.