Work on support for FMA for PlaneWaveTurbulence
This changes the way that cmake flags related to SIMD work,
both to allow manually enabling FMA and to allow detecting
it automatically with -march=native. It also adds some extra
define checks to PlaneWaveTurbulence, since -march=native
makes it impossible for cmake to determine whether the
necessary extensions are enabled. This means that we have
to check them in PlaneWaveTurbulence instead.